refactor(farm): 按功能模块拆分文件，日志降级为 Debug
- 拆分为 9 个文件：main / config / types / lifecycle / event / store / command / view / util
- slog.Info 统一降级为 slog.Debug（非必要插件）
  重点：OnEvent 里 "收到消息" / "处理群聊命令" 每条消息都打，污染最严重
- 移除 ensureDefaults，默认值由 main 提供（与 setu 一致）
